pyproject.toml
[project]
name = "tvde-qr"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = ["fastapi (>=0.128.0,<0.129.0)", "uvicorn (>=0.40.0,<0.41.0)", "jinja2 (>=3.1.6,<4.0.0)", "python-multipart (>=0.0.22,<0.0.23)", "sqlalchemy (>=2.0.46,<3.0.0)", "alembic (>=1.18.1,<2.0.0)", "psycopg[binary] (>=3.3.2,<4.0.0)", "pydantic-settings (>=2.12.0,<3.0.0)", "httpx (>=0.28.1,<0.29.0)"]

[tool.poetry]
packages = [{ include = "tvde_qr", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "respx (>=0.22.0,<0.23.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]
